pyproject.toml
[project]
name = "mlops-hw1"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "datasets>=4.4.2",
    "html5lib>=1.1",
    "huggingface>=0.0.1",
    "hydra-core>=1.3.2",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "isort>=7.0.0",
    "jsonlines>=4.0.0",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "mlflow>=3.8.1",
    "omegaconf>=2.3.0",
    "opencv-python>=4.12.0.88",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "shortuuid>=1.0.13",
    "tensorboard>=2.20.0",
    "timm>=1.0.24",
    "torch>=2.9.1",
    "torch-model-archiver>=0.12.0",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "dvc>=3.66.1",
    "pytest>=9.0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
